fix(ant): give the SDM footpod its own clock, with sub-second resolution
Page 1's time field came from treadmill.elapsed_s, which is not a footpod
clock. ble_central.c's iFit path assigned it a literal 0 on every
notification (iFit frames carry no elapsed time), and ftms_parse.c fills it
only when the treadmill sets Treadmill Data flag bit 10, memset-ing it to 0
otherwise. So on an iFit treadmill the footpod broadcast a frozen clock for
a whole run while its distance field advanced normally — an odometer moving
against a stopped clock, and a division by zero for any receiver computing
speed or stride rate from it.

23856353712_ACTIVITY.fit (48 min, 2026-08-05) shows the damage: all 8 laps
recorded total_distance 0.0 and carried no avg_speed at all, while
record.distance climbed normally to 8814.56 m.

ant_sdm.c already had correct RTC integration — it was fenced inside the
SDM:TGT branch. Lift it out so it runs in both modes and with no treadmill
connected, wrapped at 256 s (where the wire field rolls anyway) to keep
float32 resolution constant over a long session.

Fill page 1's fractional time byte too. It was independently hard-zero,
which matters for the same reason: the footpod broadcasts at ~4 Hz, so a
whole-second clock repeats its timestamp across three of every four pages
and dt reads 0 on those. Simulated over 1200 pages across the 256 s roll,
zero consecutive pages now share a time field.

treadmill_state_t.elapsed_s becomes a float to carry the fraction, and the
misleading elapsed_s = 0 assignment in on_hvx_ifit() is gone.

This does NOT fix cadence — the stride counter still never moves, so the
watch still records a flat 0 spm. Evidence and the remaining options are in
docs/sdm-recording-analysis.md; HANDOFF open item 5 is answered negatively.

+++ core/model.h
@@ -5,5 +5,11 @@ typedef struct {
     float speed_mps;
     float distance_m;
     float incline_pct;
-    uint32_t elapsed_s;
+    /* Seconds, fractional. Float rather than whole seconds because the SDM
+     * page-1 time field has a 1/256 s fractional byte and the footpod
+     * broadcasts at ~4 Hz: at whole-second resolution three of every four
+     * pages repeat the same timestamp, so a receiver differentiating against
+     * this clock divides by zero three times a second. See
+     * core/ant_sdm_encode.c ant_sdm_encode_page1(). */
+    float elapsed_s;
 } treadmill_state_t;
